revert: go back to old api and refactor description props
