+++ packages/core/src/a11y/index.ts
@@ -1,2 +1,3 @@
 export * from './useErrorMessage';
 export * from './useLabel';
+export * from './useDescription';
