+++ packages/core/src/useFormField/useFormField.ts
@@ -1,17 +1,6 @@
 import { computed, inject, MaybeRefOrGetter, nextTick, readonly, Ref, shallowRef, toValue, watch } from 'vue';
 import { FormContext, FormKey } from '../useForm/useForm';
-import {
-  AriaDescribableProps,
-  AriaDescriptionProps,
-  AriaErrorMessageProps,
-  AriaLabelableProps,
-  AriaLabelProps,
-  Arrayable,
-  Getter,
-  Maybe,
-  StandardSchema,
-  ValidationResult,
-} from '../types';
+import { Arrayable, Getter, StandardSchema, ValidationResult } from '../types';
 import { useSyncModel } from '../reactivity/useModelSync';
 import {
   cloneDeep,
@@ -21,12 +10,10 @@ import {
   tryOnScopeDispose,
   warn,
   isLowPriority,
-  createDescribedByProps,
 } from '../utils/common';
 import { FormGroupKey } from '../useFormGroup';
 import { usePathPrefixer } from '../helpers/usePathPrefixer';
 import { createDisabledContext } from '../helpers/createDisabledContext';
-import { ErrorableAttributes, useErrorMessage, useLabel } from '../a11y';
 
 interface FormFieldOptions<TValue = unknown> {
   path: MaybeRefOrGetter<string | undefined> | undefined;
@@ -37,12 +24,6 @@ interface FormFieldOptions<TValue = unknown> {
   modelName: string;
   disabled: MaybeRefOrGetter<boolean | undefined>;
   schema: StandardSchema<TValue>;
-}
-
-interface FormFieldInit {
-  inputId: MaybeRefOrGetter<string>;
-  label: MaybeRefOrGetter<string>;
-  description?: MaybeRefOrGetter<string | undefined>;
 }
 
 export type FormField<TValue> = {
@@ -56,12 +37,6 @@ export type FormField<TValue> = {
   submitErrors: Ref<string[]>;
   submitErrorMessage: Ref<string | undefined>;
   schema: StandardSchema<TValue> | undefined;
-  descriptionProps: Ref<AriaDescriptionProps>;
-  describedByProps: Ref<AriaDescribableProps>;
-  accessibleErrorProps: Ref<ErrorableAttributes>;
-  errorMessageProps: Ref<AriaErrorMessageProps>;
-  labelProps: Ref<AriaLabelProps>;
-  labelledByProps: Ref<AriaLabelableProps>;
   validate(mutate?: boolean): Promise<ValidationResult>;
   getPath: Getter<string | undefined>;
   getName: Getter<string | undefined>;
@@ -72,18 +47,10 @@ export type FormField<TValue> = {
   form?: FormContext | null;
 };
 
-interface FormControlContext {
-  inputEl: Ref<Maybe<HTMLElement>>;
-}
-
-export function useFormField<TValue = unknown>(
-  init: FormFieldInit,
-  opts?: Partial<FormFieldOptions<TValue>>,
-): FormField<TValue> {
+export function useFormField<TValue = unknown>(opts?: Partial<FormFieldOptions<TValue>>): FormField<TValue> {
   const form = inject(FormKey, null);
   const formGroup = inject(FormGroupKey, null);
   const pathPrefixer = usePathPrefixer();
-  const controlContext = shallowRef<FormControlContext>();
   const isDisabled = createDisabledContext(opts?.disabled);
   const getPath = () => {
     const path = toValue(opts?.path);
@@ -154,22 +121,6 @@ export function useFormField<TValue = unknown>(
     });
   }
 
-  const { descriptionProps, describedByProps } = createDescribedByProps({
-    inputId: init.inputId,
-    description: init.description,
-  });
-
-  const { accessibleErrorProps, errorMessageProps } = useErrorMessage({
-    inputId: init.inputId,
-    errorMessage: errorMessage,
-  });
-
-  const { labelProps, labelledByProps } = useLabel({
-    for: init.inputId,
-    label: init.label,
-    targetRef: () => controlContext.value?.inputEl.value,
-  });
-
   const field: FormField<TValue> = {
     fieldValue: readonly(fieldValue) as Ref<TValue | undefined>,
     isTouched: readonly(isTouched) as Ref<boolean>,
@@ -188,12 +139,6 @@ export function useFormField<TValue = unknown>(
     registerControl,
     submitErrors,
     submitErrorMessage,
-    descriptionProps,
-    describedByProps,
-    accessibleErrorProps,
-    errorMessageProps,
-    labelProps,
-    labelledByProps,
   };
 
   if (!form) {
@@ -260,10 +205,6 @@ export function useFormField<TValue = unknown>(
 
     form.setFieldDisabled(path, disabled);
   });
-
-  function registerControl(ctx: FormControlContext) {
-    controlContext.value = ctx;
-  }
 
   return { ...field, form };
 }
@@ -528,21 +469,6 @@ export type ExposedField<TValue> = {
    * @param mutate - Whether to set errors on the field as a result of the validation call, defaults to `true`.
    */
   validate: (mutate?: boolean) => Promise<ValidationResult>;
-
-  /**
-   * Props for the description element.
-   */
-  descriptionProps: Ref<AriaDescriptionProps>;
-
-  /**
-   * Props for the error message element.
-   */
-  errorMessageProps: Ref<AriaErrorMessageProps>;
-
-  /**
-   * Props for the label element.
-   */
-  labelProps: Ref<AriaLabelProps>;
 };
 
 export function exposeField<TReturns extends object, TValue>(
@@ -571,9 +497,6 @@ export function exposeField<TReturns extends object, TValue>(
     setTouched: field.setTouched,
     setValue: field.setValue,
     validate: (mutate = true) => field.validate(mutate),
-    descriptionProps: field.descriptionProps,
-    errorMessageProps: field.errorMessageProps,
-    labelProps: field.labelProps,
   } satisfies ExposedField<TValue>;
 
   return {
